document update

+++ UltraEmeraldScriptEditor/EditorSupport/Undo/UndoStack.cs
@@ -15,9 +15,16 @@ namespace EditorSupport.Undo
 
         public virtual void AddOperation(IUndoableOperation operation)
         {
-            _operations.RemoveRange(_undoProcess, _operations.Count - _undoProcess);
-            _operations.Add(operation);
-            ++_undoProcess;
+            if (_group != null)
+            {
+                _group.Operations.Add(operation);
+            }
+            else
+            {
+                _operations.RemoveRange(_undoProcess, _operations.Count - _undoProcess);
+                _operations.Add(operation);
+                ++_undoProcess;
+            }
         }
 
         public virtual Boolean Undo()
@@ -58,7 +65,28 @@ namespace EditorSupport.Undo
             return _undoProcess < _operations.Count;
         }
 
+        public void StartGrouping()
+        {
+            if (_group != null)
+            {
+                throw new InvalidOperationException("UndoStack has already started.");
+            }
+            _group = new UndoOperationGroup();
+        }
+
+        public void EndGrouping()
+        {
+            if (_group == null)
+            {
+                throw new InvalidOperationException("UndoStack hasn't started yet.");
+            }
+            var group = _group;
+            _group = null;
+            AddOperation(group);
+        }
+
         protected List<IUndoableOperation> _operations;
+        private UndoOperationGroup _group;
         protected Int32 _undoProcess;
     }
 }
